Simplify sibling window detection
Simplify sibling window detection by assuming that another vertically
maximized window on the same monitor is the sibling. I'm not sure why I
thought it was necessary before to also check whether the windows are
horizontally tiled.

This fixes an issue where a valid sibling was sometimes not detected
because `rightRect.x` was one pixel short.